move model properties object to registry dir
