+++ src/marqo_commons/model_registry/model_properties_object.py
